pyproject.toml
[project]
name = "hello-langgraph-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "ipython>=9.6.0,<10.0",
    "langchain[openai]>=1.0.3,<2.0",
    "langgraph>=1.0.2,<2.0",
    "pillow>=12.0.0,<13.0",
    "pydantic>=2.12.3,<3.0",
    "python-dotenv>=1.2.1,<2.0",
]

[project.optional-dependencies]
cli = [
    "langgraph-cli[inmem]>=0.4.6,<0.5.0",
    "pyright>=1.1.407",
    "ruff>=0.14.3,<0.15.0",
]


[tool.pyright]
venvPath = "."
venv = ".venv"
defineConstant = { DEBUG = true }
typeCheckingMode = "strict"
include = ["src"]
strict = ["src"]
# stubPath = "src/stubs" # TODO?
# ignore = ["src/oldstuff"] # NOTE what is the difference between exclude and ignore?
exclude = [
    ".venv",
    "typings",
    "**/__pycache__",
    "**/test*.py",
    "**/*.as-is.py"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/hello_langgraph"
]
